Fixed structure and ready to make builds
