+++ .gitignore
@@ -9,8 +9,8 @@ undo.json
 /.idea/vcs.xml
 /ZIPPER_2_COVER_SETTER_UI.ui
 /junk_stuff/
-/MangaManager/CoverManagerLib/.pytest_cache/
-/MangaManager/MetadataManagerLib/.pytest_cache/
+/MangaManager/src/CoverManagerLib/.pytest_cache/
+/MangaManager/src/MetadataManagerLib/.pytest_cache/
 /MangaManager/tests/.pytest_cache/
 /MangaManager/tests/Sample CBZ Chapter 23.cbz
 /MangaManager/tests/Sample CBZ Chapter 24.cbz
