+++ MangaManager/src/CommonLib/HelperFunctions.py
