+++ MangaManager/src/CommonLib/SettingsWindow.py
