+++ MangaManager/src/ConvertersLib/epub2cbz/epub2cbz.py
