+++ MangaManager/src/ConvertersLib/epub2cbz/epub2cbz.ui
@@ -1,0 +1,82 @@
+<?xml version='1.0' encoding='utf-8'?>
+<interface version="1.2">
+    <object class="tk.Frame" id="frame_1">
+        <property name="height">200</property>
+        <property name="padx">50</property>
+        <property name="pady">50</property>
+        <property name="width">200</property>
+        <layout manager="grid">
+            <property name="column">0</property>
+            <property name="row">0</property>
+        </layout>
+        <containerlayout manager="grid">
+            <property type="row" id="2" name="pad">20</property>
+        </containerlayout>
+        <child>
+            <object class="tk.Label" id="label_1">
+                <property name="font">{Title} 20 {}</property>
+                <property name="text" translatable="yes">EPUB 2 CBZ</property>
+                <layout manager="grid">
+                    <property name="column">0</property>
+                    <property name="row">0</property>
+                </layout>
+            </object>
+        </child>
+        <child>
+            <object class="tk.Label" id="label_2">
+                <property name="font">{SUBTITLE} 12 {}</property>
+                <property name="text" translatable="yes">This script extracts the images from a .epub file to a .cbz
+                    file.
+                </property>
+                <layout manager="grid">
+                    <property name="column">0</property>
+                    <property name="row">1</property>
+                </layout>
+            </object>
+        </child>
+        <child>
+            <object class="tk.Button" id="button_1">
+                <property name="command" type="command" cbtype="simple">_select_files</property>
+                <property name="text" translatable="yes">Load .epub file</property>
+                <layout manager="grid">
+                    <property name="column">0</property>
+                    <property name="row">2</property>
+                </layout>
+            </object>
+        </child>
+        <child>
+            <object class="tk.Label" id="label_3">
+                <property name="text" translatable="yes">Selected files:</property>
+                <layout manager="grid">
+                    <property name="column">0</property>
+                    <property name="row">3</property>
+                </layout>
+            </object>
+        </child>
+        <child>
+            <object class="tk.Listbox" id="listbox_1">
+                <property name="activestyle">dotbox</property>
+                <property name="font">{courier} 12 {}</property>
+                <property name="justify">center</property>
+                <property name="width">69</property>
+                <layout manager="grid">
+                    <property name="column">0</property>
+                    <property name="row">4</property>
+                </layout>
+            </object>
+        </child>
+        <child>
+            <object class="tk.Button" id="button_2">
+                <property name="command" type="command" cbtype="simple">_processFile</property>
+                <property name="compound">top</property>
+                <property name="cursor">based_arrow_down</property>
+                <property name="font">TkDefaultFont</property>
+                <property name="text" translatable="yes">Process</property>
+                <layout manager="grid">
+                    <property name="column">0</property>
+                    <property name="row">5</property>
+                </layout>
+            </object>
+        </child>
+    </object>
+</interface>
