+++ MangaManager/src/CoverManagerLib/CoverDownloader.ui
