+++ MangaManager/src/CoverManagerLib/DELETE_COVER.jpg
