+++ MangaManager/src/CoverManagerLib/RECOVER_COVER.jpg
