+++ MangaManager/src/CoverManagerLib/errors.py
