+++ MangaManager/src/CoverManagerLib/models.py
