+++ MangaManager/src/MetadataManagerLib/ComicInfo.py
