+++ MangaManager/src/MetadataManagerLib/MetadataManager_UI.ui
