+++ MangaManager/src/VolumeManager/VolumeManager_UI.ui
