+++ MangaManager/src/VolumeManager/errors.py
