+++ MangaManager/src/VolumeManager/models.py
